moves tests into test folder & cleans up some old code
