+++ src/test/ViewCharger.test.js
